Added seg_seg_coll_dir() case for parallel vectors.
The average vector in that case is just equal to both of them,
and special logic is needed. I'm hoping this holds up for
horizontal and vertical cases, so this function can be used
for all seg-seg normal resolutions. Then it can be factored
out, which could be good because for polygons made of many
segments, the normal direction might only be appropriate
to compute when it's already known which segments are forming
colliding corners (if any).